requirements.txt
asgiref==3.8.1
click==8.1.7
coverage==7.5.1
crispy-tailwind==1.0.3
Django==5.0.4
django-allauth==0.62.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-htmx==1.17.3
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg==3.1.19
pytest==8.2.1
pytest-django==4.8.0
sqlparse==0.5.0
typing_extensions==4.12.0
uvicorn==0.29.0
